First commit of work in progress for dividing up
Right now this just has copies of all of the templates in the CloudFormation
repo, split off into their own module directory and with the variables that
the template requires reformatted into a variables.tf file.  The root dir is
made from the stack.yaml file, and also includes its defaults in the
variables.tf.

We might not want to make these all modules, or might want to use
community modules instead, but starting with this format helps make it
easier to start chopping up the work into discrete portions for Erin and I
to share.