Add conversion of securitygroups to terraform
There are two TODOs I'm not sure about.  Otherwise I think this is
good.  This isn't meant to last, but is just me taking a full blast at
converting all the security groups at once.  When we do new modules,
we can pull these over to the relevant modules and delete from here,
so that we can be sure we're not forgetting any security groups that
we need.  Eventually the file should be empty and we can just delete
the entire directory.